#103 refine readme and test
